+++ creme/ensemble/base.py
@@ -1,0 +1,7 @@
+import collections
+
+from .. import base
+
+
+class Ensemble(base.Estimator, collections.UserList):
+    pass